Refactor component imports for consistency; create index files for easier access and update type definitions
