+++ components/hero/index.ts
@@ -1,0 +1,7 @@
+/**
+ * @author ColdByDefault
+ * @copyright 2025 ColdByDefault. All Rights Reserved.
+ */
+
+export { default as Hero } from "./Hero";
+export { default as SocialLinks } from "./SocialLinks";
